move interpolate function code from modules to functional
